requirement.txt
boto
python-digitalocean
ansible
